Apply smart-crop to post images
Currently we only hook into `tachyon_image_downsize_string` which is a Tachyon hook that fires on `image_downsize` when using functions like `wp_get_attachment_image_src`. When Tachyon converts image URLs in the post content to Tachyon'ed images, no such function is called. We need to make use of the `tachyon_post_image_args` to filter those. The same args are passed to both filters, so it's a simple enough fix.