Add full image size props when returning Media Lib response
When the Media Library fetches the thumbnails, the full size was missing because full is not a registered size in WP.

This change prevents a JS error when selecting an image to insert in Gutenberg (but could potentially appear somewhere else). As a result, the Image Size selector wasn't working (though this commit only solveds it partially).