Fixed Image Size Selector in Image Block
The Image Size selector wasn't working because the Tachyon URL wasn't properly calculated. The tachyon_url() function was receiving already a Tachyon URL so it was returned as it is without any change.

This was happening when the Image Block fetches for the image data by using the attachment REST API endpoint.