Fix cropper showing on Featured Media Gutenberg popup
Sinse WordPress 5.3, the Feature Image media modal has not been showing the Cropper UI. After many trials and tribulations, I the closest commit responsible in Core I found was https://github.com/WordPress/WordPress/commit/03390f00c58a5d1a21052397e414ee43be845dd3. Not all that helpful, as this is a large sync from (effectively) Gutenberg. Looking into the Gutenberg commit history, it looks like roughly when https://github.com/WordPress/gutenberg/commit/e62a670e400506472f36b70a5cc90913552d5633 was committed, the custom Cropper UI of Smart media stopped working for the Feature Image modal.

Long story (kinda) short is that Gutenberg introduced it's own `MediaFrame` view subclass called `FeaturedImage` that is instantiated by the Gutenberg react component. This means the `state` of the Media frame is instead called `featured-image` (from `library` in the normal case). Cropper makes a few assumptions that the media frame will always have a state called `library` (of which `featured-image` is a subclass). So though it should all mostly work, the change in ID breaks several things from working.

We instead need to support both possible states of the frame. There are some other small differences too (like a different set of `this.options`) which need to be accounted for, which have been in this commit.

Due to the private API natiure of this whole endeavor, there's a good amount of trying to battle the code to get the "edit" state (that's the Cropper state) to override the intended one from Core. Specifically we have introduced another workaround when setting the active state to re-trigger the `activate` event.

This should be backwards compatible with WordPress 5.2, as well as tested in 5.3.