Prevent srcset requesting images larger than the original
In Tachyon we removed the ability to upscale images beyond the original dimensions, as a result the zoom modifiers could request a larger image than the original but not receive the expected size back from Tachyon.

Because the `srcset` modifier didnt account for this the browser would incorrectly render the image at a smaller size than intended because it assumed it was dealing with a larger image. This resulted in a shrinking effect when rendered in the browser.

This ensures that `srcset` URLs are never larger than the original image.

Additionally it seems in some browsers if there any `srcset` URLs available it will never fall back to the original `img` tag URL but instead use the closest `srcset` URL it can find, because of this we also force a `srcset` size with `zoom=1` to ensure the original size is available as well.