Fix metabox compat
Adds a focusin event listener to the post editor to check for focus events on metaboxes to deselect the currently selected block. This will prevent selecting an image using the media modal updating a currently selected block.

Note this approach is not complete or ideal as there are performance implications, focus events happen regularly inside the editor.