Check uploader status is rendered
In some situations the uploader status is no longer rendered but has not been disposed properly before the `info` method is called which tries to update DOM elements that have not been set yet. The specific case this occurs in is when using the uploader view that then switches back to the library view.