Fix fake attachment handling
In some instances the block editor provides default images such as the cover block. These do not exist as real attachments but are represented by an attachment object containing just an ID property. This updates checks for additional attachment object data before proceeding to switch to the edit state.

Fixes #108