Workaround AMP's media library modifications
AMP makes some alterations to the media modal at a later stage that breaks Smart Media's edit mode. The change is just that it adds a notice to the media library with the recommended image size but this is repeated in the sidebar.

A future TODO will be to do a deep dive into the root cause and make this properly compatible.